TRUNK GS: - added weight limit for pets (need DP update, 'pet_stats' table) - fixed life time/food bar for summons/pets - fixed bug with PC weight penalty status update - fixed pet food rate option in config - added/reworked some system messages
TODO: 
    - pet inventory slots
    - pet weight penalty

git-svn-id: http://svn.l2jfree.com/l2jfree@904 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/L2_GameServer_c5/src/main/java/net/sf/l2j/Config.java
@@ -528,7 +528,7 @@ public final class Config {
     
 	/** Pet Rates (Multipliers) */ 	
 	public static float         PET_XP_RATE; 
- 	public static int           PET_FOOD_RATE;
+ 	public static float         PET_FOOD_RATE;
 
     /** Karma Drop Rate control */
     public static int   KARMA_DROP_LIMIT;
@@ -1697,7 +1697,7 @@ public final class Config {
                 PLAYER_RATE_DROP_EQUIP_WEAPON   = Integer.parseInt(ratesSettings.getProperty("PlayerRateDropEquipWeapon", "5"));
 
                 PET_XP_RATE                     = Float.parseFloat(ratesSettings.getProperty("PetXpRate", "1."));
-                PET_FOOD_RATE                   = Integer.parseInt(ratesSettings.getProperty("PetFoodRate", "1")); 
+                PET_FOOD_RATE                   = Float.parseFloat(ratesSettings.getProperty("PetFoodRate", "1")); 
 
                 KARMA_DROP_LIMIT                = Integer.parseInt(ratesSettings.getProperty("KarmaDropLimit", "10"));
                 KARMA_RATE_DROP                 = Integer.parseInt(ratesSettings.getProperty("KarmaRateDrop", "70"));
